Do not allow environment overrides with default value.
as @zmughal, empty string '' unsets an environment variable on
Windows, and thus the default I had chosen was unwise as it did
subtly different things on different platforms.  If someone really
does want to set to '' they still can, but they have to explicitly
do it.  Also we note that it may be unportable in the documentation.