Don't break system utilities on use of bins property
When alien_version_check is set, we may end up depending on the
system-installed version of the built package. In that case, the
absolute path in the dist_dir won't be available, and the binaries fail
to work.

Check for that using the ConfigData package, and use the binary in
/usr/bin directlry, assuming it's installed there.